keep selected pokemon on text input change
